requirements.txt
requests==2.25.1
aiohttp==3.7.3
alembic==1.4.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
bleach==3.2.1
blinker==1.4
boto3==1.16.36
botocore==1.19.36
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cssselect==1.1.0
decorator==4.4.2
dicttoxml==1.7.4
docutils==0.16
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
future==0.18.2
fuzzywuzzy==0.18.0
geocoder==1.38.1
geographiclib==1.50
geoip2==4.1.0
geopy==2.0.0
idna==2.10
ip2geotools==0.1.5
IP2Location==8.5.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
keyring==21.8.0
lazy-object-proxy==1.4.3
lxml==4.6.2
Mako==1.1.3
MarkupSafe==1.1.1
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
packaging==20.8
pip-review==1.1.0
pkginfo==1.6.1
psycopg2
pycodestyle==2.6.0
Pygments==2.7.3
PyJWT==1.7.1
pylint==2.6.0
pylint-flask-sqlalchemy
pylint-flask
pyparsing==2.4.7
PyPDF2==1.26.0
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-Levenshtein
pywin32-ctypes==0.2.0
ratelim==0.1.6
readme-renderer==28.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Searchable==1.2.0
SQLAlchemy-Utils==0.36.8
toml==0.10.2
tqdm==4.56.0
twine==3.3.0
typed-ast==1.4.2
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.2
validators==0.18.1
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
python-telegram-bot
gunicorn
